+++ sdk/peering/arm-peering/src/operations/index.ts
@@ -22,3 +22,5 @@ export * from "./peeringServiceLocations";
 export * from "./prefixes";
 export * from "./peeringServiceProviders";
 export * from "./peeringServices";
+export * from "./resourceMove";
+export * from "./rpUnbilledPrefixes";