+++ sdk/peering/arm-peering/src/operations/peeringServiceLocations.ts
@@ -144,7 +144,6 @@ const listNextOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.ErrorResponse
     }
   },
-  queryParameters: [Parameters.apiVersion, Parameters.country],
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,