+++ sdk/peering/arm-peering/src/operations/prefixes.ts
@@ -336,7 +336,6 @@ const listByPeeringServiceNextOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.ErrorResponse
     }
   },
-  queryParameters: [Parameters.apiVersion, Parameters.expand],
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,