+++ sdk/peering/arm-peering/src/peeringManagementClient.ts
@@ -30,7 +30,9 @@ import {
   PeeringServiceLocationsImpl,
   PrefixesImpl,
   PeeringServiceProvidersImpl,
-  PeeringServicesImpl
+  PeeringServicesImpl,
+  ResourceMoveImpl,
+  RpUnbilledPrefixesImpl
 } from "./operations";
 import {
   CdnPeeringPrefixes,
@@ -48,7 +50,9 @@ import {
   PeeringServiceLocations,
   Prefixes,
   PeeringServiceProviders,
-  PeeringServices
+  PeeringServices,
+  ResourceMove,
+  RpUnbilledPrefixes
 } from "./operationsInterfaces";
 import * as Parameters from "./models/parameters";
 import * as Mappers from "./models/mappers";
@@ -91,7 +95,7 @@ export class PeeringManagementClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-peering/2.1.1`;
+    const packageDetails = `azsdk-js-arm-peering/3.0.0`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -144,7 +148,7 @@ export class PeeringManagementClient extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2021-06-01";
+    this.apiVersion = options.apiVersion || "2023-08-01";
     this.cdnPeeringPrefixes = new CdnPeeringPrefixesImpl(this);
     this.legacyPeerings = new LegacyPeeringsImpl(this);
     this.lookingGlass = new LookingGlassImpl(this);
@@ -161,6 +165,8 @@ export class PeeringManagementClient extends coreClient.ServiceClient {
     this.prefixes = new PrefixesImpl(this);
     this.peeringServiceProviders = new PeeringServiceProvidersImpl(this);
     this.peeringServices = new PeeringServicesImpl(this);
+    this.resourceMove = new ResourceMoveImpl(this);
+    this.rpUnbilledPrefixes = new RpUnbilledPrefixesImpl(this);
     this.addCustomApiVersionPolicy(options.apiVersion);
   }
 
@@ -224,6 +230,8 @@ export class PeeringManagementClient extends coreClient.ServiceClient {
   prefixes: Prefixes;
   peeringServiceProviders: PeeringServiceProviders;
   peeringServices: PeeringServices;
+  resourceMove: ResourceMove;
+  rpUnbilledPrefixes: RpUnbilledPrefixes;
 }
 // Operation Specifications
 const serializer = coreClient.createSerializer(Mappers, /* isXml */ false);